% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-deployment.R, R/mod-detection-path.R,
%   R/mod-detection-ratio.R
\name{mod_deployment_server}
\alias{mod_deployment_server}
\alias{mod_detection_path_ui}
\alias{mod_detection_path_server}
\alias{mod_detection_ratio_server}
\title{PDF Download Module}
\usage{
mod_deployment_server(id, detection, deployment, station, river)

mod_detection_path_ui(id)

mod_detection_path_server(id)

mod_detection_ratio_server(id)
}
\description{
A shiny module for downloading field data cards. Select and
download resources needed for the project like field cards.
}
\keyword{internal}
