% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-detection-ratio.R
\name{mod_detection_ratio_ui}
\alias{mod_detection_ratio_ui}
\title{PDF Download Module}
\usage{
mod_detection_ratio_ui(id)
}
\description{
A shiny module for downloading field data cards. Select and
download resources needed for the project like field cards.
}
\keyword{internal}
