% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-deployment.R
\name{mod_deployment_ui}
\alias{mod_deployment_ui}
\title{PDF Download Module}
\usage{
mod_deployment_ui(id)
}
\description{
A shiny module for downloading field data cards. Select and
download resources needed for the project like field cards.
}
\keyword{internal}
