% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-detection-path.R, R/mod-detection-ratio.R
\name{mod_detection_path_ui}
\alias{mod_detection_path_ui}
\alias{mod_detection_path_server}
\alias{mod_detection_ratio_server}
\title{Detection path}
\usage{
mod_detection_path_ui(id, sex, weight, forklength)

mod_detection_path_server(
  id,
  detection_event,
  deployment,
  fish,
  reference_rkm,
  lims_x,
  lims_y,
  receiver_group,
  receiver_group_rkm
)

mod_detection_ratio_server(id, detection_complete, fish, receiver_group)
}
\description{
A shiny module for visualizing detection paths.
}
\keyword{internal}
