% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-detection-ratio.R
\name{mod_detection_ratio_ui}
\alias{mod_detection_ratio_ui}
\title{Detection ratio}
\usage{
mod_detection_ratio_ui(id, sex, weight, forklength)
}
\description{
A shiny module for visualizing absolute and relative detection ratios.
}
\keyword{internal}
