% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(
  detection_station = readRDS(system.file("extdata/detection_station.rds", package =
    "shinywsgcolr")),
  detection_event = readRDS(system.file("extdata/detection_event.rds", package =
    "shinywsgcolr")),
  detection_complete = readRDS(system.file("extdata/detection_complete.rds", package =
    "shinywsgcolr")),
  deployment = readRDS(system.file("extdata/deployment.rds", package = "shinywsgcolr")),
  fish = readRDS(system.file("extdata/fish.rds", package = "shinywsgcolr")),
  station = readRDS(system.file("extdata/station.rds", package = "shinywsgcolr")),
  reference_rkm = readRDS(system.file("extdata/reference_rkm.rds", package =
    "shinywsgcolr")),
  river = readRDS(system.file("extdata/river.rds", package = "shinywsgcolr")),
  timestep = "day",
  receiver_group = "array",
  receiver_group_rkm = "array_rkm"
)
}
\arguments{
\item{detection_station}{A tibble of detection station data returned from `wsgcolr::detection_timestep()` with `receiver_group` of `station_id`. Mandatory columns include 'transmitter_id', 'timestep', and 'station_id'.}

\item{detection_event}{A tibble of detection path data returned from \code{wsgcolr::detection_event()}containing columns 'transmitter_id', 'timestep', 'path', 'array', 'array_rkm'.}

\item{detection_complete}{A tibble of complete detection data returned from \code{wsgcolr::detection_complete()} containing columns 'transmitter_id' and 'timestep'.}

\item{deployment}{A tibble of the receiver deployments containing columns 'date_deployment', 'date_last_download', 'station', 'array', 'array_rkm'.}

\item{fish}{A tibble of fish data containing columns 'transmitter_id', 'sex', 'forklength_cm', 'weight_kg'.}

\item{station}{A sf object of the station point locations containing column 'array' (for color coding). Geometry column must inherit class 'sfc_POINT' or 'sfc_MULTIPOINT'.}

\item{reference_rkm}{A tibble of reference rkm locations containing columns 'label' and 'rkm'.}

\item{river}{A sf object of the river polygon. Geometry column must inherit class 'sfc_POLYGON' or 'sfc_MULTIPOLYGON'.}

\item{timestep}{A character string of the timestep. Acceptable values are those accepted by \code{lubridate::floor_date()}: second, minute, hour, day, week, month, bimonth, quarter, season, halfyear and year.}

\item{receiver_group}{A character string of the column containing receiver group (e.g., station, zone, array).}

\item{receiver_group_rkm}{A string of the column name indicating receiver_group rkm.}
}
\description{
Run the Shiny Application
}
